// crm-ui/src/app/api.generated.service.ts
export interface Company {
    id: number;
    name?: string | null;
    address?: string | null;
}